Moving remaining tools out of old imageTools folder
